+++ packages/html/src/icon/svg-icon.tsx
@@ -15,9 +15,7 @@ const options = {
         ThemeColor.success,
         ThemeColor.warning,
         ThemeColor.error,
-        ThemeColor.info,
-        ThemeColor.light,
-        ThemeColor.dark
+        ThemeColor.info
     ],
 };