+++ packages/html/src/segmented-control/tests/segmented-control.tsx
@@ -1,5 +1,4 @@
-import { SegmentedControlNormal } from '../../segmented-control';
-
+import { SegmentedControlNormal } from "../../segmented-control";
 
 const styles = `
     #test-area>section {
@@ -9,35 +8,26 @@ const styles = `
         gap: 10px;
         justify-items: start;
     }
-
-    .k-segmented-control-thumb {
-        width: 100px;
-    }
-
-    .full-width .k-segmented-control-thumb {
-        width: 300px;
-    }
 `;
 
 export default () => (
-    <>
-        <style>{styles}</style>
-        <div id="test-area">
-            <section>
-                <div>Default</div>
-                <SegmentedControlNormal />
-            </section>
+  <>
+    <style>{styles}</style>
+    <div id="test-area">
+      <section>
+        <div>Default</div>
+        <SegmentedControlNormal thumbStyles={{ width: "50%" }} />
+      </section>
 
-            <section style={{ width: "200px"}}>
-                <div>Stretched (small width) - items with equal widths</div>
-                <SegmentedControlNormal stretched />
-            </section>
+      <section style={{ width: "200px" }}>
+        <div>Stretched (small width) - items with equal widths</div>
+        <SegmentedControlNormal thumbStyles={{ width: "50%" }} stretched />
+      </section>
 
-            <section>
-                <div>Stretched - 100% width - items with equal widths</div>
-                <SegmentedControlNormal stretched className="full-width" />
-            </section>
-        </div>
-
-    </>
+      <section>
+        <div>Stretched - 100% width - items with equal widths</div>
+        <SegmentedControlNormal thumbStyles={{ width: "50%" }} stretched className="full-width" />
+      </section>
+    </div>
+  </>
 );